Part 3 of writing the NUANMB importer
I did not pay attention to the fact that the data needs to be read in
another buffer, so after fixing my code to take that into account,
animation data is being read correctly. However, compressed data is not
yet fully supported.

Major changes:
* The animation import script now has more functions for reading different
kinds of data; one for direct, and another for compressed
* Buffer1 and Buffer2 have been renamed to BufferOffset and BufferSize,
respectively, to better reflect their purpose; they are used as part of
the buffer used to read the actual animation data

Minor changes:
* The AnimTrack class no longer contains multiple arrays; only one array
is defined now