The data read-only scripts now depend on Blender (but not on its UI)
This dependency has been added so that the scripts have access to mathutils
and all of its functions. As a side effect, matrices and vectors are printed
in a slightly nicer-looking format.

To run these scripts, type this into a terminal window/command prompt:
blender --background  --python <path-to-script>, where `blender` may need to
be replaced by the full executable path depending on how Blender was installed.

Minor changes:
* Add a zero to the position and scale vectors to represent W