Part 5 of writing the NUANMB importer
The animation data is getting imported into Blender, but most of it is
unfortunately incorrect! However, not all is really lost, but I will continue
to find out what went wrong during the import process. The data is being read
correctly, but I may be missing something else that I did not pay attention to.

Major changes:
* The animation name, frame count, the list of bones to be animated, and
visibility animations are correctly applied; note that visibility animations
have an effect while in Blender that could be lost if models are exported

Minor changes (for both importers):
* Timeline markers are changed to become pose-specific markers
* Fake users are set for all actions created