Part 11 of writing the NUANMB importer (really for the NUMDLB importer)
Finally! After several days of struggling, the NUMDLB importer has been fixed so
that animations will deform imported models correctly. This could potentially
save people from having to learn how to do animations unless they want to.

Major changes:
* Bones are transformed first, so that axises can be aligned
* Bone length is calculated based on the armature size, instead of always
pointing up on the Y-axis
* Bone matrices are set before the bone length is set, with bone roll preserved

Minor changes:
None, as the focus of this commit is to fix bone axises, for which work is done